implemented construction location javascript. fixes #21

+++ src/test/java/de/tum/bgu/msm/models/javascript/LeaveParentHhTest.java
@@ -22,7 +22,7 @@ public class LeaveParentHhTest {
     @Test
     public void testModelOne() throws ScriptException {
         calculator.setPersonType(31);
-        Assert.assertEquals(0.0003, (double) calculator.calculate(), 0.);
+        Assert.assertEquals(0.0003, calculator.calculate(), 0.);
     }
 
     @Test(expected = ScriptException.class)
